pyproject.toml
[project]
name = "mem0-tcvectordb"
version = "0.1.0b0"
description = "TCVectorDB backend for the mem0 memory framework."
readme = "README.md"
requires-python = ">=3.9,<4.0"
authors = [
    { name = "Alvie Zhang" },
]
license = { file = "LICENSE" }
keywords = ["mem0", "memory", "vector database", "tencent cloud", "tcvectordb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "mem0ai>=1.0.0,<2.0.0",
    "tcvectordb>=1.8.5",
    "rank-bm25>=0.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/alviezhang/mem0-tcvectordb"
Source = "https://github.com/alviezhang/mem0-tcvectordb"
Issues = "https://github.com/alviezhang/mem0-tcvectordb/issues"

[tool.hatch.build.targets.wheel]
packages = ["mem0_tcvectordb"]

[tool.hatch.build.targets.sdist]
include = [
    "mem0_tcvectordb",
    "pyproject.toml",
    "LICENSE",
    "README.md",
]
ignore-vcs = true

[tool.hatch.build]
exclude = [
    "__pycache__",
    "*.pyc",
    "uv.lock",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--cov=mem0_tcvectordb", "--cov-report=term", "--cov-report=xml:coverage.xml"]
pythonpath = ["."]
markers = [
    "integration: marks tests that exercise the external Mem0+TCVectorDB pipeline",
]

[dependency-groups]
dev = [
    "langchain-neo4j>=0.5.0",
    "neo4j>=5.0.0",
    "python-dotenv>=1.0.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.5",
]
